feat(healthyme-payment): reportes
- logica para generar reportes de diferentes areas
- logica para exportar los datos en un archivo pdf

+++ healthyme-payment/src/main/java/studio/devbyjose/healthyme_payment/controller/PagoController.java
@@ -68,7 +68,7 @@ public class PagoController {
     }
 
     @GetMapping("/paciente/{idPaciente}")
-    public ResponseEntity<List<PagoDTO>> getPagosByPaciente(@PathVariable Integer idPaciente) {
+    public ResponseEntity<List<PagoDTO>> getPagosByPaciente(@PathVariable Long idPaciente) {
         List<PagoDTO> pagos = pagoService.getPagosByPaciente(idPaciente);
         return ResponseEntity.ok(pagos);
     }
